Move vocabularies to config/ directory
