+++ src/components/Browse/BrowseFilterTool.vue
@@ -68,7 +68,7 @@
 import { computed, ref, watch } from 'vue'
 import { Ref } from 'vue'
 import { Thing } from '@/types'
-import { siteTypes } from '@/vocabularies'
+import { siteTypes } from '@/config/vocabularies'
 import { useDisplay } from 'vuetify/lib/framework.mjs'
 
 const { smAndDown } = useDisplay()
