+++ src/components/Metadata/SensorFormCard.vue
@@ -87,7 +87,7 @@ import { VForm } from 'vuetify/components'
 import { useFormLogic } from '@/composables/useFormLogic'
 import { rules } from '@/utils/rules'
 import { computed } from 'vue'
-import { methodTypes } from '@/vocabularies'
+import { methodTypes } from '@/config/vocabularies'
 import { Sensor } from '@/types'
 
 const props = defineProps({ sensor: Object as () => Sensor })
