+++ src/components/Site/SiteForm.vue
@@ -180,7 +180,7 @@ import { onMounted, reactive, ref, watch } from 'vue'
 import GoogleMap from '../GoogleMap.vue'
 import { useThingStore } from '@/store/thing'
 import { Thing } from '@/types'
-import { siteTypes } from '@/vocabularies'
+import { siteTypes } from '@/config/vocabularies'
 import { VForm } from 'vuetify/components'
 import { rules } from '@/utils/rules'
 import { storeToRefs } from 'pinia'
