+++ src/config/vocabularies.ts
